requirements.txt
hydra-core==1.3.2
matplotlib==3.6.3
nibabel==5.1.0
numpy==1.23.3
omegaconf==2.3.0
opencv_python_headless==4.8.1.78
pandas==1.5.1
Pillow==10.1.0
pytorch_lightning==2.1.1
scikit_learn==1.3.2
torch==2.0.1
torchmetrics==1.2.0
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.64.1
